Enable model caching (#576)
Switches the VSCode extension to use the new CacheManager from Malloy.

Truncated cache works the same way it did before, except that for dependencies of the main file, it too will use the "real" cache.

We used to remove documents from the cache when the file was closed. Now we check if the file is used by any of the other files in the cache first. If A depends on B, closing B will not cause it to be removed from the cache, but then closing A will clear both. If you close A first, it will be removed from the cache, since nothing open depends on it.